backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
flask-bcrypt==1.0.1
flask-jwt-extended==4.5.2 